Introduce AdaptiveGraphManager
