+++ flink-runtime/src/main/java/org/apache/flink/runtime/jobgraph/IntermediateDataSet.java
@@ -19,6 +19,7 @@
 package org.apache.flink.runtime.jobgraph;
 
 import org.apache.flink.runtime.io.network.partition.ResultPartitionType;
+import org.apache.flink.streaming.api.graph.StreamEdge;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -42,6 +43,8 @@ public class IntermediateDataSet implements java.io.Serializable {
 
     // All consumers must have the same partitioner and parallelism
     private final List<JobEdge> consumers = new ArrayList<>();
+
+    private final List<StreamEdge> streamEdges = new ArrayList<>();
 
     // The type of partition to use at runtime
     private final ResultPartitionType resultType;
@@ -85,6 +88,10 @@ public class IntermediateDataSet implements java.io.Serializable {
         return resultType;
     }
 
+    public List<StreamEdge> getConsumerStreamEdges() {
+        return streamEdges;
+    }
+
     // --------------------------------------------------------------------------------------------
 
     public void addConsumer(JobEdge edge) {
@@ -103,6 +110,22 @@ public class IntermediateDataSet implements java.io.Serializable {
         consumers.add(edge);
     }
 
+    public void addStreamEdge(StreamEdge edge) {
+        DistributionPattern pattern =
+                edge.getPartitioner().isPointwise()
+                        ? DistributionPattern.POINTWISE
+                        : DistributionPattern.ALL_TO_ALL;
+        if (streamEdges.isEmpty()) {
+            distributionPattern = pattern;
+            isBroadcast = edge.getPartitioner().isBroadcast();
+        } else {
+            checkState(distributionPattern == pattern, "Incompatible distribution pattern.");
+            checkState(
+                    isBroadcast == edge.getPartitioner().isBroadcast(),
+                    "Incompatible broadcast type.");
+        }
+        streamEdges.add(edge);
+    }
     // --------------------------------------------------------------------------------------------
 
     @Override
