+++ flink-runtime/src/main/java/org/apache/flink/streaming/api/graph/StreamEdge.java
@@ -238,4 +238,8 @@ public class StreamEdge implements Serializable {
     public IntermediateDataSetID getIntermediateDatasetIdToProduce() {
         return intermediateDatasetIdToProduce;
     }
+
+    public String getId() {
+        return edgeId;
+    }
 }
